#Merge Payment Ref with Ecotech laundry
