--- pos_payment_ref/models/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import pos_payment
-from . import pos_order
