--- pos_payment_ref/models/pos_payment.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-from odoo import fields, models
-
-
-class PosPayment(models.Model):
-    _inherit = 'pos.payment'
-
-    payment_ref = fields.Char('Payment Reference')
-
-class PosPayment(models.Model):
-    _inherit = 'pos.payment.method'
-
-    pos_payment_ref = fields.Boolean('Recored Payment Ref', default=False)
