--- pos_payment_ref/static/src/css/screens.css
@@ -1,7 +1,0 @@
-
-.pos .paymentline.selected .input-ref{
-width:7em;
-}
-.pos .paymentline .input-ref{
-width:7em;
-}
