--- pos_payment_ref/views/pos_order_view.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="pos_paymentref_pos_order_view_form" model="ir.ui.view">
-        <field name="name">pos.payment.pos.order.form</field>
-        <field name="model">pos.order</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_pos_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field//tree/field[@name='payment_method_id']" position="after">
-                <field name="payment_ref" string="Payment Ref"/>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="view_pos_payment_form_inherot" model="ir.ui.view">
-        <field name="name">pos.payment.inherit.form</field>
-        <field name="model">pos.payment</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_payment_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='payment_method_id']" position="after">
-                <field name="payment_ref" string="Payment Ref"/>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="view_pos_payment_tree_inherot" model="ir.ui.view">
-        <field name="name">pos.payment.inherit.tree</field>
-        <field name="model">pos.payment</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_payment_tree"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='payment_method_id']" position="after">
-                <field name="payment_ref" string="Payment Ref"/>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="pos_payment_method_view_form_inherit" model="ir.ui.view">
-        <field name="name">pos.payment.method.inherit.tree</field>
-        <field name="model">pos.payment.method</field>
-        <field name="inherit_id" ref="point_of_sale.pos_payment_method_view_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='company_id']" position="after">
-                <field name="pos_payment_ref"/>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
