--- pos_payment_ref/views/pos_payment_ref.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <template id="assets_backend" name="pos_payment_ref assets" inherit_id="point_of_sale.assets">
-        <xpath expr="." position="inside">
-            <link rel="stylesheet" href="/pos_payment_ref/static/src/css/screens.css"/>
-            <script type="text/javascript" src="/pos_payment_ref/static/src/js/models.js"></script>
-            <script type="text/javascript" src="/pos_payment_ref/static/src/js/popups.js"></script>
-            <script type="text/javascript" src="/pos_payment_ref/static/src/js/screens.js"></script>
-        </xpath>
-    </template>
-</odoo>
-
